Fix: Move ValueChanged to event tag 16 and carry one f64 (FR-15.2.5)
The picker implementation put the value change event on tag 8, which FR-18
reserves for the pointer gestures, and gave it an i64 payload. FR-15.2.5 says
tag 16 and one f64: integers up to 2^53 survive an f64 exactly, so an epoch
day count, a slider position and a toggle all fit one event rather than
growing a second name for the same concept.

Covers the FR-15.2.5 criteria that the record is tag 16, 24 bytes, and that
the reserved tags 8 to 15 do not decode. User code is unchanged: DatePicker
still takes EventHandler<i64> and converts in the component.

+++ dioxus-compose-renderer/desktop/src/renderer/HostDropdown.kt
@@ -74,7 +74,7 @@ internal fun HostDropdown(
             selected = index
             val handlerId = node.handler(PropertyKind.OnValueChange)
             if (handlerId != null) {
-                dispatcher.dispatch(HostEvent.ValueChanged(node.id, handlerId, index.toLong()))
+                dispatcher.dispatch(HostEvent.ValueChanged(node.id, handlerId, index.toDouble()))
             }
         }
     }
